Restructured project structure
