+++ src/Data/Repo/iTestRepo.java
@@ -1,0 +1,6 @@
+package Data.Repo;
+
+import Data.Repo.iRepo;
+import Model.Test;
+
+public interface iTestRepo extends iRepo<Test, Integer> { }
